rename fdesc to fmsg and remove fault.Msg entirely
